// src/types/case-study.ts
export interface CaseStudyType {
    id: string;
    title: string;
    tags: string[];
    serviceType: string;
    shortDescription: string;
    urlSlug: string;
    thumbnailUrl: string;
    description: string;
    metaDescription: string;
    active: boolean;
    authorId: string;
    createdAt: string;
    updatedAt: string;
    urlBanner:string
}